Deep hygiene: AmbientSound PLLength, identifying-PC types, bools

+++ src/wiz8/local_screens/RCSItemsPage.cpp
@@ -872,7 +872,7 @@ unsigned char BackpackRegionHandler005BB350(const InputAtom* event, W8Region* re
     if (event->usEvent == 0x100) {
         if ((region->flags & W8_REGION_RIGHT_BUTTON_HELD) != 0 && item->item_id != -1 &&
             g_camp_screen_0069c0f4->entry_mode != 3) {
-            g_camp_entry_parameter_0069c0fc = (int)g_value_0069c0f8;
+            g_camp_entry_parameter_0069c0fc = g_value_0069c0f8;
             if (CanItemLeaveItsSlot(item) != 0 && PartyAttemptsToIdentifyItem(item, 0) != 0 &&
                 g_camp_screen_0069c0f4->realm_flags[1] != 0) {
                 RebuildCampItemList005A4A00();
@@ -994,7 +994,7 @@ unsigned char EquipSlotRegionHandler005BB560(const InputAtom* event, W8Region* r
         }
         if ((region->flags & W8_REGION_RIGHT_BUTTON_HELD) != 0 && item->item_id != -1 &&
             g_camp_screen_0069c0f4->entry_mode != 3) {
-            g_camp_entry_parameter_0069c0fc = (int)g_value_0069c0f8;
+            g_camp_entry_parameter_0069c0fc = g_value_0069c0f8;
             if (CanItemLeaveItsSlot(item) != 0 && PartyAttemptsToIdentifyItem(item, 0) != 0 &&
                 g_camp_screen_0069c0f4->realm_flags[1] != 0) {
                 RebuildCampItemList005A4A00();
@@ -1034,7 +1034,7 @@ unsigned char ItemPoolRegionHandler005BB900(const InputAtom* event, W8Region* re
             if ((region->flags & W8_REGION_RIGHT_BUTTON_HELD) != 0 &&
                 pool_index < (unsigned int)g_status_685170.party_item_count_1791 &&
                 g_camp_screen_0069c0f4->entry_mode != 3) {
-                g_camp_entry_parameter_0069c0fc = (int)g_value_0069c0f8;
+                g_camp_entry_parameter_0069c0fc = g_value_0069c0f8;
                 if (CanItemLeaveItsSlot(item) != 0 && PartyAttemptsToIdentifyItem(item, 0) != 0 &&
                     g_camp_screen_0069c0f4->realm_flags[1] != 0) {
                     RebuildCampItemList005A4A00();
